requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
alembic
pydantic
httpx
python-dotenv
APScheduler
pytest
pytest-asyncio